+++ bugsnag-android-core/src/main/java/com/bugsnag/android/ManifestConfigLoader.kt
@@ -39,6 +39,7 @@ internal class ManifestConfigLoader {
         private const val MAX_PERSISTED_SESSIONS = "$BUGSNAG_NS.MAX_PERSISTED_SESSIONS"
         private const val LAUNCH_CRASH_THRESHOLD_MS = "$BUGSNAG_NS.LAUNCH_CRASH_THRESHOLD_MS"
         private const val LAUNCH_DURATION_MILLIS = "$BUGSNAG_NS.LAUNCH_DURATION_MILLIS"
+        private const val SEND_LAUNCH_CRASHES_SYNCHRONOUSLY = "$BUGSNAG_NS.SEND_LAUNCH_CRASHES_SYNCHRONOUSLY"
         private const val APP_TYPE = "$BUGSNAG_NS.APP_TYPE"
     }
 
@@ -84,6 +85,10 @@ internal class ManifestConfigLoader {
                     LAUNCH_DURATION_MILLIS,
                     launchDurationMillis.toInt()
                 ).toLong()
+                sendLaunchCrashesSynchronously = data.getBoolean(
+                    SEND_LAUNCH_CRASHES_SYNCHRONOUSLY,
+                    sendLaunchCrashesSynchronously
+                )
             }
         }
         return config