+++ bugsnag-android-core/src/test/java/com/bugsnag/android/InternalEventPayloadDelegateTest.kt
@@ -38,6 +38,7 @@ internal class InternalEventPayloadDelegateTest {
         `when`(this.appDataCollector.generateAppWithState()).thenReturn(app)
         app.durationInForeground = 500L
         app.inForeground = true
+        app.isLaunching = true
         `when`(
             deviceDataCollector
                 .generateDeviceWithState(ArgumentMatchers.anyLong())
@@ -65,6 +66,7 @@ internal class InternalEventPayloadDelegateTest {
         // app
         assertEquals(500L, event.app.durationInForeground)
         assertEquals(true, event.app.inForeground)
+        assertEquals(true, event.app.isLaunching)
         assertNotNull(event.app)
 
         // device