+++ CHANGELOG.md
@@ -5,7 +5,12 @@
 ### Deprecations
 
 * `Configuration#launchCrashThresholdMs` is deprecated in favour of `Configuration#launchDurationMillis`
+
+### Enhancements
+
+* Add public API for crash-on-launch detection
   [#1157](https://github.com/bugsnag/bugsnag-android/pull/1157)
+  [#1159](https://github.com/bugsnag/bugsnag-android/pull/1159)
 
 ## 5.7.0 (2021-02-18)